% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdtmf.R
\name{export_resdtmf}
\alias{export_resdtmf}
\title{Export a DFM into resdtmf}
\usage{
export_resdtmf(input_dfm, file_prefix, compress = FALSE)
}
\arguments{
\item{input_dfm}{dfm object}

\item{file_prefix}{characters, file prefix of the 3 exported files}

\item{compress}{logical, to compress the 3 files into a zip file}
}
\description{
This function exports a dfm into the Responsible Document-term Matrix format.
}
