% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdtmf.R
\name{export_resdtmf}
\alias{export_resdtmf}
\title{Export a DFM into resdtmf}
\usage{
export_resdtmf(input_dfm, file_path, compress = FALSE)
}
\arguments{
\item{input_dfm}{dfm object}

\item{file_path}{characters, file path of the exported file}

\item{compress}{logical, to compress the 3 files into a zip file (not implemented yet)}
}
\description{
This function exports a dfm into the Responsible Document-term Matrix format.
}
