% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdtmf.R
\name{import_resdtmf}
\alias{import_resdtmf}
\title{Import resdtmf files into a dfm}
\usage{
import_resdtmf(file_path)
}
\arguments{
\item{file_path}{characters, file path of the json file.}
}
\description{
This function imports resdtmf files exported using export_resdtmf into a dfm object.
}
