% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdtmf.R
\name{import_resdtmf}
\alias{import_resdtmf}
\title{Import a resdtmf file into DFM}
\usage{
import_resdtmf(file_path)
}
\arguments{
\item{file_path}{characters, file path of the resdtmf json file.}
}
\description{
This function imports a resdtmf file exported using export_resdtmf into a dfm object.
}
